+++ zingolib/src/test_framework/mocks.rs
@@ -142,7 +142,7 @@ mod sapling_crypto_note {
             let (_, _, address) = default_zaddr();
             Self::new()
                 .recipient(address)
-                .value(NoteValue::from_raw(1000000))
+                .value(NoteValue::from_raw(200000))
                 .rseed(Rseed::AfterZip212([7; 32]))
         }
     }
@@ -177,7 +177,7 @@ pub mod orchard_note {
         let fvk: FullViewingKey = (&sk).into();
         let recipient = fvk.address_at(0u32, Scope::External);
 
-        let value = NoteValue::default(); // ZERO
+        let value = NoteValue::from_raw(800000);
         let rho = {
             loop {
                 let mut bytes = [0u8; 32];