+++ zingolib/src/wallet/describe.rs
@@ -12,7 +12,7 @@ use zcash_note_encryption::Domain;
 use zcash_primitives::consensus::BlockHeight;
 
 use crate::wallet::data::TransactionRecord;
-use crate::wallet::notes::NoteInterface;
+use crate::wallet::notes::OutputInterface;
 use crate::wallet::notes::ShieldedNoteInterface;
 
 use crate::wallet::traits::Diversifiable as _;
@@ -70,7 +70,7 @@ impl LightWallet {
                     filtered_notes
                         .map(|notedata| {
                             if notedata.spent().is_none() && notedata.pending_spent().is_none() {
-                                <D::WalletNote as ShieldedNoteInterface>::value(notedata)
+                                <D::WalletNote as OutputInterface>::value(notedata)
                             } else {
                                 0
                             }